Refactor dto
